espresso_contract_deployer::proposals

Module multisig

Structs

LightClientV2UpgradeParams

Parameters for upgrading LightClient to V2

StakeTableV2UpgradeParams

TransferOwnershipParams

Enums

MultisigOwnerCheck

Functions

encode_generic_calldata

Encode calldata for any function call.

encode_transfer_ownership_calldata

Encode transferOwnership(address) calldata.

encode_upgrade_calldata

Encode upgradeToAndCall(address,bytes) calldata for a proxy upgrade.

transfer_ownership_from_multisig_to_timelock

upgrade_esp_token_v2_multisig_owner

Upgrade the EspToken proxy to use EspTokenV2. Deploys new implementation, then returns encoded upgrade calldata.

upgrade_fee_contract_multisig_owner

Upgrade the FeeContract proxy to a new implementation (patch upgrade). Deploys new implementation, then returns encoded upgrade calldata.

upgrade_light_client_v2_multisig_owner

Upgrade the light client proxy to use LightClientV2. Deploys new implementation contracts, then returns encoded upgrade calldata.

upgrade_light_client_v3_multisig_owner

Upgrade the light client proxy to use LightClientV3. Deploys new implementation contracts, then returns encoded upgrade calldata.

upgrade_stake_table_v2_multisig_owner

Upgrade the stake table proxy to use StakeTableV2. Deploys new implementation and returns encoded upgrade calldata.