espresso_contract_deployer::proposals::multisig

Function upgrade_esp_token_v2_multisig_owner

pub async fn upgrade_esp_token_v2_multisig_owner(
    provider: impl Provider,
    contracts: &mut Contracts,
    rpc_url: String,
    dry_run: Option<bool>,
) -> Result<(String, bool)>

Upgrade the EspToken proxy to use EspTokenV2. Internally, first detect existence of proxy, then deploy EspTokenV2, then upgrade and initializeV2. Assumes:

the proxy is already deployed.

the proxy is owned by a multisig.

Returns the url link to the upgrade proposal This function can only be called on a real network supported by the safeSDK