espresso_contract_deployer::proposals::multisig

Function upgrade_light_client_v3_multisig_owner

pub async fn upgrade_light_client_v3_multisig_owner(
    provider: impl Provider,
    contracts: &mut Contracts,
    is_mock: bool,
    rpc_url: String,
    dry_run: Option<bool>,
) -> Result<String>

Upgrade the light client proxy to use LightClientV3. Internally, first detect existence of proxy, then deploy LCV3, then upgrade and initializeV3. Internal to “deploy LCV3”, we deploy PlonkVerifierV3 whose address will be used at LCV3 init time. Assumes:

the proxy is already deployed.

the proxy is owned by a multisig.

the proxy is not yet initialized for V3

Returns the url link to the upgrade proposal This function can only be called on a real network supported by the safeSDK