espresso_contract_deployer::proposals

Module multisig

Structs

LightClientV2UpgradeParams

Parameters for upgrading LightClient to V2

TransferOwnershipParams

Functions

call_transfer_ownership_script

Call the transfer ownership script to transfer ownership of a contract to a new owner

call_upgrade_proxy_script

Call the upgrade proxy script to upgrade a proxy contract

find_script_path

transfer_ownership_from_multisig_to_timelock

upgrade_esp_token_v2_multisig_owner

Upgrade the EspToken proxy to use EspTokenV2. Internally, first detect existence of proxy, then deploy EspTokenV2, then upgrade and initializeV2. Assumes:

upgrade_light_client_v2_multisig_owner

Upgrade the light client proxy to use LightClientV2. Internally, first detect existence of proxy, then deploy LCV2, then upgrade and initializeV2. Internal to “deploy LCV2”, we deploy PlonkVerifierV2 whose address will be used at LCV2 init time. Assumes:

upgrade_stake_table_v2_multisig_owner

Upgrade the stake table proxy to use StakeTableV2. Internally, first detect existence of proxy, then deploy StakeTableV2 Assumes:

verify_node_js_files

Verify the node js files are present and can be executed.